requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.0.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
black==23.12.1
blinker==1.6.2
blis==0.7.11
build==1.0.3
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.3
cloudpathlib==0.16.0
cmake==3.27.7
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.4
ctransformers==0.2.27
cymem==2.0.8
dataclasses-json==0.5.14
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.7
entrypoints==0.4
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.108.0
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.3.3
google-api-core==2.15.0
google-auth==2.26.1
google-generativeai==0.2.2
googleapis-common-protos==1.62.0
greenlet==2.0.2
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==6.1.1
InstructorEmbedding==1.0.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
kubernetes==29.0.0
langchain==0.0.348
langchain-community==0.0.13
langchain-core==0.1.11
langchain-google-genai==0.0.6
langchainplus-sdk==0.0.10
langcodes==3.3.0
langsmith==0.0.77
lit==17.0.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnxruntime==1.16.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
pandas==2.0.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.0.0
posthog==3.3.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.7
pulsar-client==3.4.0
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pypdf==3.17.4
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
semver==3.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.16
srsly==2.4.8
starlette==0.32.0.post1
streamlit==1.25.0
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.3
thinc==8.2.1
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
triton==2.0.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.1.0
uvicorn==0.26.0
uvloop==0.19.0
validators==0.20.0
virtualenv==20.24.7
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.21.0
weasel==0.3.4
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.2
zipp==3.15.0
